+++ functional-tests/functional/dart/test/SkipElement_test.dart
@@ -28,9 +28,6 @@ final _testSuite = TestSuite("SkipElement");
 class SkipTagsInDartImpl implements SkipTagsInDart {
   @override
   void dontSkipTagged() {}
-
-  @override
-  void release() {}
 }
 
 void main() {