+++ gluecodium/src/test/resources/smoke/listeners/output/dart/lib/src/smoke/listener_with_properties.dart
@@ -39,9 +39,7 @@ abstract class ListenerWithProperties {
     bufferedMessageGetLambda,
     bufferedMessageSetLambda
   );
-  /// @nodoc
-  @Deprecated("Does nothing")
-  void release() {}
+
   String get message;
   set message(String value);
   CalculationResult get packedMessage;
@@ -226,8 +224,7 @@ class ListenerWithProperties$Lambdas implements ListenerWithProperties {
     this.bufferedMessageGetLambda,
     this.bufferedMessageSetLambda
   );
-  @override
-  void release() {}
+
   @override
   String get message => messageGetLambda();
   @override
@@ -259,8 +256,7 @@ class ListenerWithProperties$Lambdas implements ListenerWithProperties {
 }
 class ListenerWithProperties$Impl extends __lib.NativeBase implements ListenerWithProperties {
   ListenerWithProperties$Impl(Pointer<Void> handle) : super(handle);
-  @override
-  void release() {}
+
   String get message {
     final _getFfi = __lib.catchArgumentError(() => __lib.nativeLibrary.lookupFunction<Pointer<Void> Function(Pointer<Void>, Int32), Pointer<Void> Function(Pointer<Void>, int)>('library_smoke_ListenerWithProperties_message_get'));
     final _handle = this.handle;